archive/api/requirements_api.txt
# API Requirements for Multi-Agent Academic Advising System
# For Railway/Docker deployment

# =============================================================================
# Core Framework
# =============================================================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# =============================================================================
# Database
# =============================================================================
motor>=3.3.0                 # Async MongoDB driver
pymongo>=4.6.0               # Sync MongoDB (for some operations)
certifi>=2023.0.0            # SSL certificates for MongoDB Atlas

# =============================================================================
# Authentication
# =============================================================================
python-jose[cryptography]>=3.3.0    # JWT tokens
passlib[bcrypt]>=1.7.4              # Password hashing
python-multipart>=0.0.6             # Form data parsing

# =============================================================================
# LangChain & AI
# =============================================================================
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.0.20
openai>=1.0.0

# =============================================================================
# Vector Database
# =============================================================================
chromadb>=0.4.0

# =============================================================================
# Utilities
# =============================================================================
python-dotenv>=1.0.0
httpx>=0.26.0
email-validator>=2.1.0

# =============================================================================
# Production
# =============================================================================
gunicorn>=21.2.0             # Production WSGI server (backup)
